Update user api
