+++ src/pages/api/users/index.ts
@@ -1,0 +1,36 @@
+import { NextApiHandler } from "next";
+
+import { prisma } from "../../../shared/prisma-helper";
+import { User } from "../../../types";
+
+export const getUsers = async (): Promise<User[] | null> => {
+  const users = await prisma.user.findMany({
+    select: {
+      name: true,
+      avatarUrl: true,
+      auth0userSub: true,
+      email: false,
+    },
+  });
+
+  return users as User[];
+};
+
+const createUser = async (newUser: User): Promise<User> => {
+  const user: User = await prisma.user.create({
+    data: newUser,
+  });
+
+  return user;
+};
+
+const handler: NextApiHandler = async (req, res) => {
+  if (req.method === "POST") {
+    const user = await createUser(JSON.parse(req.body));
+    res.status(200).json({ user });
+  } else if (req.method === "GET") {
+    res.status(200).json({ user: await getUsers() });
+  }
+};
+
+export default handler;
